requirements.txt
streamlit==1.35.0
tensorflow==2.15.0
keras
librosa==0.10.0.post2
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
pillow==9.5.0
